Add usage data and license info to all data source docs
Revise all data source descriptions to follow a common pattern:
1.) What, in brief, is it?
2.) Why is it useful to include in Mapzen Search?
3.) What is the license?

I'd love some further help with tweaks in a couple areas that I think
are still lacking:
* keeping it informative without sounding too sales-pitchy. I really
want to list the good reasons why we use each data source without making
it sound like we are trying to sell someone something they don't need.
* Following the common format without sounding too templated
* Not hating on Geonames. It's a good dataset and we include it for a
reason, but it is a bit different and we also want to capture that
honestly.